+++ rds-20140815/alibabacloud_rds20140815/models/_create_rcimage_request.py
@@ -1,0 +1,61 @@
+# -*- coding: utf-8 -*-
+# This file is auto-generated, don't edit it. Thanks.
+from __future__ import annotations
+
+from darabonba.model import DaraModel
+
+class CreateRCImageRequest(DaraModel):
+    def __init__(
+        self,
+        image_name: str = None,
+        instance_id: str = None,
+        region_id: str = None,
+        snapshot_id: str = None,
+    ):
+        # The name of the custom image.
+        self.image_name = image_name
+        # The ID of the RDS Custom instance.
+        self.instance_id = instance_id
+        # The region ID. You can call the DescribeRegions operation to query the most recent region list.
+        self.region_id = region_id
+        # The ID of the snapshot from which to create the custom image. You can call the DescribeRCSnapshots operation to query the snapshot ID.
+        self.snapshot_id = snapshot_id
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.image_name is not None:
+            result['ImageName'] = self.image_name
+
+        if self.instance_id is not None:
+            result['InstanceId'] = self.instance_id
+
+        if self.region_id is not None:
+            result['RegionId'] = self.region_id
+
+        if self.snapshot_id is not None:
+            result['SnapshotId'] = self.snapshot_id
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('ImageName') is not None:
+            self.image_name = m.get('ImageName')
+
+        if m.get('InstanceId') is not None:
+            self.instance_id = m.get('InstanceId')
+
+        if m.get('RegionId') is not None:
+            self.region_id = m.get('RegionId')
+
+        if m.get('SnapshotId') is not None:
+            self.snapshot_id = m.get('SnapshotId')
+
+        return self
+